vanna_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
vanna[chromadb,openai,postgres]
psycopg[binary]
python-dotenv
